feat: remove otel test code
